Update Cocoapods to 7.11.0 and C++ SDK version to 7.3.0. (#395)
* Update Cocoapods to 7.11.0.

* Update readme file Cocoapods version.

* Update C++ SDK version to 7.3.0.

* Add release notes to readme.

* Update Firestore source dependency to 7.11.0.

* Fix a compiler warning for size_t/int.

* Update iOS integration test OS deployment target to 10.0.

* Pin googletest to release 1.10.0, as master is now broken on iOS.

+++ testing/sample_framework/src/ios/ios_app_framework.mm
@@ -184,7 +184,7 @@ void *stdout_logger(void *filedes_ptr) {
   int fd = reinterpret_cast<int *>(filedes_ptr)[0];
   std::string buffer;
   char bufchar;
-  while (int n = read(fd, &bufchar, 1)) {
+  while (size_t n = read(fd, &bufchar, 1)) {
     if (bufchar == '\0') {
       break;
     } else if (bufchar == '\n') {
